Implemented automatically building climb when csv file of ticks is imported

+++ RouteTickrAPI/src/DTOs/ClimbDto.cs
@@ -9,7 +9,7 @@ public class ClimbDto
     public string Rating { get; set; } = string.Empty;
     public string Location { get; set; } = string.Empty;
     public string Url { get; set; } = string.Empty;
-    public decimal? Height { get; set; }
+    public double? Height { get; set; }
     public ClimbDangerRating? DangerRating { get; set; }
     public ClimbType ClimbType { get; set; }
